[Refactor] : ArticleDetailCard -> ArticleDetail 로 컴포넌트 이름 수정
