--- src/components/ArticleDetail/index.tsx
@@ -1,2 +1,0 @@
-export { default as ArticleDetailCard } from './ArticleDetailCard/index';
-export { default as ProgressBar } from './ProgressBar/index';
