+++ src/components/ArticleView/ProgressBar/ProgressBarContent.tsx
