+++ src/components/ArticleView/ProgressBar/index.tsx
