+++ src/components/ArticleView/ProgressBar/style.ts
