+++ src/components/ArticleView/index.tsx
@@ -1,0 +1,2 @@
+export { default as ArticleDetail } from './ArticleDetail/index';
+export { default as ProgressBar } from './ProgressBar/index';
